Re-factor the AWS IAM module into its own directory.
This new directory structure lets the individual actors get imported
into the main actors.aws.iam namespace, allowing each one to be in a
smaller and more managable file.
