+++ kingpin/actors/aws/iam/base.py
@@ -1,0 +1,40 @@
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Copyright 2014 Nextdoor.com, Inc
+
+"""
+:mod:`kingpin.actors.aws.iam.base`
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+"""
+
+import logging
+
+from tornado import concurrent
+
+from kingpin.actors.aws import base
+
+log = logging.getLogger(__name__)
+
+__author__ = 'Mikhail Simin <mikhail@nextdoor.com>'
+
+
+# This executor is used by the tornado.concurrent.run_on_executor()
+# decorator. We would like this to be a class variable so its shared
+# across RightScale objects, but we see testing IO errors when we
+# do this.
+EXECUTOR = concurrent.futures.ThreadPoolExecutor(10)
+
+
+class IAMBaseActor(base.AWSBaseActor):
+
+    """Base class for IAM actors."""
