+++ kingpin/actors/aws/iam/entities.py
@@ -13,8 +13,8 @@
 # Copyright 2014 Nextdoor.com, Inc
 
 """
-:mod:`kingpin.actors.aws.iam`
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+:mod:`kingpin.actors.aws.iam.entities`
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 """
 
 import json
@@ -25,14 +25,14 @@ from boto.exception import BotoServerError
 from tornado import concurrent
 from tornado import gen
 
-from kingpin.actors.aws import base
 from kingpin.actors import exceptions
+from kingpin.actors.aws.iam import base
 from kingpin.constants import REQUIRED
 from kingpin.constants import STATE
 
 log = logging.getLogger(__name__)
 
-__author__ = 'Mikhail Simin <mikhail@nextdoor.com>'
+__author__ = 'Matt Wise <matt@nextdoor.com>'
 
 
 # This executor is used by the tornado.concurrent.run_on_executor()
@@ -42,161 +42,16 @@ __author__ = 'Mikhail Simin <mikhail@nextdoor.com>'
 EXECUTOR = concurrent.futures.ThreadPoolExecutor(10)
 
 
-class IAMBaseActor(base.AWSBaseActor):
-
-    """Base class for IAM actors."""
-
-
-class UploadCert(IAMBaseActor):
-
-    """Uploads a new SSL Cert to AWS IAM.
-
-    **Options**
-
-    :private_key_path:
-      (str) Path to the private key.
-
-    :path:
-      (str) The AWS "path" for the server certificate. Default: "/"
-
-    :public_key_path:
-      (str) Path to the public key certificate.
-
-    :name:
-      (str) The name for the server certificate.
-
-    :cert_chain_path:
-      (str) Path to the certificate chain. Optional.
-
-    **Example**
-
-    .. code-block:: json
-
-       { "actor": "aws.iam.UploadCert",
-         "desc": "Upload a new cert",
-         "options": {
-           "name": "new-cert",
-           "private_key_path": "/cert.key",
-           "public_key_path": "/cert.pem",
-           "cert_chain_path": "/cert-chain.pem"
-         }
-       }
-
-    **Dry run**
-
-    Checks that the passed file paths are valid. In the future will also
-    validate that the files are of correct format and content.
-    """
-
-    all_options = {
-        'name': (str, REQUIRED, 'The name for the server certificate.'),
-        'public_key_path': (str, REQUIRED,
-                            'Path to the public key certificate.'),
-        'private_key_path': (str, REQUIRED, 'Path to the private key.'),
-        'cert_chain_path': (str, None, 'Path to the certificate chain.'),
-        'path': (str, None, 'The path for the server certificate.')
-    }
-
-    @gen.coroutine
-    def _upload(self, cert_name, cert_body, private_key, cert_chain, path):
-        """Create a new server certificate in AWS IAM."""
-        yield self.thread(
-            self.iam_conn.upload_server_cert,
-            cert_name=cert_name,
-            cert_body=cert_body,
-            private_key=private_key,
-            cert_chain=cert_chain,
-            path=path)
-
-    @gen.coroutine
-    def _execute(self):
-        """Gather all the cert data and upload it.
-
-        The `boto` library requires actual cert contents, but this actor
-        expects paths to files.
-        """
-        # Gather needed cert data
-        cert_chain_body = None
-        if self.option('cert_chain_path'):
-            cert_chain_body = self.readfile(self.option('cert_chain_path'))
-
-        cert_body = self.readfile(self.option('public_key_path'))
-        private_key = self.readfile(self.option('private_key_path'))
-
-        # Upload it
-        if self._dry:
-            self.log.info('Would upload cert "%s"' % self.option('name'))
-            raise gen.Return()
-
-        self.log.info('Uploading cert "%s"' % self.option('name'))
-        yield self._upload(
-            cert_name=self.option('name'),
-            cert_body=cert_body,
-            private_key=private_key,
-            cert_chain=cert_chain_body,
-            path=self.option('path'))
-
-
-class DeleteCert(IAMBaseActor):
-
-    """Delete an existing SSL Cert in AWS IAM.
-
-    **Options**
-
-    :name:
-      (str) The name for the server certificate.
-
-    **Example**
-
-    .. code-block:: json
-
-       { "actor": "aws.iam.DeleteCert",
-         "desc": "Run DeleteCert",
-         "options": {
-           "name": "fill-in"
-         }
-       }
-
-    **Dry run**
-
-    Will find the cert by name or raise an exception if it's not found.
-    """
-
-    all_options = {
-        'name': (str, REQUIRED, 'The name for the server certificate.')
-    }
-
-    @gen.coroutine
-    def _find_cert(self, name):
-        """Find a cert by name."""
-
-        self.log.debug('Searching for cert "%s"...' % name)
-        try:
-            yield self.thread(self.iam_conn.get_server_certificate, name)
-        except BotoServerError as e:
-            raise exceptions.UnrecoverableActorFailure(
-                'Could not find cert %s. Reason: %s' % (name, e))
-
-    @gen.coroutine
-    def _delete(self, cert_name):
-        """Delete a server certificate in AWS IAM."""
-        yield self.thread(self.iam_conn.delete_server_cert, cert_name)
-
-    @gen.coroutine
-    def _execute(self):
-        if self._dry:
-            self.log.info('Checking that the cert exists...')
-            yield self._find_cert(self.option('name'))
-            self.log.info('Would delete cert "%s"' % self.option('name'))
-            raise gen.Return()
-
-        self.log.info('Deleting cert "%s"' % self.option('name'))
-        yield self._delete(cert_name=self.option('name'))
-
-
-class User(IAMBaseActor):
+class User(base.IAMBaseActor):
 
     """Manages an IAM User.
+
+    This actor manages the state of an Amazon IAM User. It ensures that the
+    user either exists or does not. It also updates any settings for the user
+    that are different from the passed in options.
+
+    At the moment you can mange the users state, its inline policies, and you
+    can purge unmanaged inline policies.
 
     **Options**
 
@@ -207,8 +62,9 @@ class User(IAMBaseActor):
       (str) Present or Absent. Default: "present"
 
     :inline_policies:
-      (array) A list of strings that point to JSON files to use as inline
-      policies. Default: []
+      (str,array) A list of strings that point to JSON files to use as inline
+      policies. You can also pass in a single inline policy as a string.
+      Default: []
 
     :inline_policies_purge:
       (bool) Whether or not to purge un-managed policies. Default: false
